usb: dwc3: usb limit to high speed.
This works around a 'adb root' and cable disconnect-connect issue when
USB OTG is connected to a superspeed port.

o This patch limits USB max speed on a superspeed port to high speed.
o A USB-C to USB-C cable will still experience a disconnect-reconnect
issue. Use a USB-A to USB-C cable.

This patch needs to be added for the AIA-O for the NUC platform
for USB3.0 proper detection.

Jira: None.
Test: Depends on patch W/A for dwc3 revision < 2.2a
      Boot the board. Check adb with USB3.0 cable.

BUG: 32940866
Change-Id: I00a888946d729078728768a5b3327abf2fda81cc
(cherry picked from commit 483946c1afce6c714d15d4ea981fdefee86fb555)
Signed-off-by: mabbas <mohamed.abbas@intel.com>
Signed-off-by: Abhilash K V <abhilash.k.v@intel.com>
Signed-off-by: karan889 <karanx.bhagoji@intel.com>