Add Support to enable xHCI DBC
xHCI compatible USB host controllers(i.e. super-speed USB3 controllers)
can be implemented with the Debug Capability(DbC). It presents a debug
device which is fully compliant with the USB framework and provides the
equivalent of a very high performance full-duplex serial link. The debug
capability operation model and registers interface are defined in 7.6.8
of the xHCI specification, revision 1.1.

The DbC debug device shares a root port with the xHCI host. By default,
the debug capability is disabled and the root port is assigned to xHCI.
When the DbC is enabled, the root port will be assigned to the DbC debug
device, and the xHCI sees nothing on this port. This implementation uses
a sysfs node named <dbc> under the xHCI device to manage the enabling
and disabling of the debug capability.This implementation also uses a
sysfs node named <dbc_func> under the xHCI device to choose either
a tty interface or a raw character interface provided by DbC.

When the debug capability is enabled, it will present a debug device
through the debug port. This debug device is fully compliant with the
USB3 framework, and it can be enumerated by a debug host on the other
end of the USB link. As soon as the debug device is configured, if the
dbc_func is set to 1, a TTY serial device named /dev/ttyDBC0 will be
created, otherwise if the dbc_func is set to 0. a Character device
named /dev/dbc_raw will be created.

One use of this link is running a login service on the debug target.
Hence it can be remote accessed by a debug host. Another use case can
probably be found in servers. It provides a peer-to-peer USB link
between two host-only machines. This provides a reasonable out-of-band
communication method between two servers.

The raw USB function exposed to the user space as a cdev can be used
for user space USB functions development.

Jira: None
Test: None

Signed-off-by: Lu Baolu <baolu.lu@linux.intel.com>
Signed-off-by: Rajaram Regupathy <rajaram.regupathy@intel.com>
Signed-off-by: K, Gururaj <gururaj.k@intel.com>